fix: merge booking and payment, add enumerate

+++ src/main/java/org/foodie_tour/modules/vnpay/dto/request/PaymentRequest.java
@@ -10,6 +10,6 @@ import lombok.experimental.FieldDefaults;
 @Builder
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class PaymentRequest {
-    int bookingId;
-    int amount;
+    Long bookingId;
+    Long amount;
 }
